Resolve "Add Cadence and QuestaSim Plugins"

See merge request incoresemi/river-framework/core-verification/river_core_plugins!6

+++ generator_plugins/aapg_plugin/aapg_plugin.py
@@ -6,34 +6,21 @@ import os
 import sys
 import pluggy
 import shutil
-import yaml
 import random
 import re
 import datetime
 import pytest
 import glob
 from river_core.log import logger
-from river_core.utils import *
+import river_core.utils as utils
 from river_core.constants import *
 import re
 
 gen_hookimpl = pluggy.HookimplMarker("generator")
 
 
-class AapgPlugin(object):
+class aapg_plugin(object):
     """ Generator hook implementation """
-
-    #@gen_hookimpl
-    #def load_config(self, isa, platform):
-    #    pwd = os.getcwd()
-
-    #    try:
-    #        isa_file, platform_file = riscv_config.check_specs(
-    #                                    isa, platform, pwd, True)
-    #    except ValidationError as msg:
-    #        #logger.error(msg)
-    #        print('error')
-    #        sys.exit(1)
 
     # creates gendir and any setup related things
     @gen_hookimpl
@@ -50,6 +37,18 @@ class AapgPlugin(object):
 
         # Check if the path exists or not
         logger.debug('AAPG Pre Gen Stage')
+        output_dir = os.path.abspath(output_dir)
+        self.name = 'aapg'
+
+        # Check if aapg is there on path
+        if shutil.which('aapg') is None:
+            logger.error('Plugin requires aapg to be installed')
+            raise SystemExit
+
+        if shutil.which('spike') is None:
+            logger.error('AAPG requires spike to be installed')
+            raise SystemExit
+
         if (os.path.isdir(output_dir)):
             logger.debug('exists')
             shutil.rmtree(output_dir, ignore_errors=True)
@@ -64,6 +63,14 @@ class AapgPlugin(object):
         # Eventually add support for riscv_config
         self.isa = spec_config['isa']
 
+        self.json_dir = output_dir + '/../.json/'
+        logger.debug(self.json_dir)
+        # Check if dir exists
+        if (os.path.isdir(self.json_dir)):
+            logger.debug(self.json_dir + ' Directory exists')
+        else:
+            os.makedirs(self.json_dir)
+
     # gets the yaml file with list of configs; test count; parallel
     @gen_hookimpl
     def gen(self, gen_config, module_dir, output_dir):
@@ -73,21 +80,23 @@ class AapgPlugin(object):
         pytest_file = module_dir + '/aapg_plugin/gen_framework.py'
         logger.debug(pytest_file)
 
-        # if norun:
-        #     # to display test items
-        #     pytest.main([pytest_file, '--collect-only', '-n={0}'.format(jobs), '-k={0}'.format(filter), '--configlist={0}'.format(gen_config), '-v',  '--seed={0}'.format(seed), '--count={0}'.format(count), '--html=aapg_gen.html', '--self-contained-html'])
-        # else:
+        output_dir = os.path.abspath(output_dir)
+
+        report_file_name = '{0}/{1}_{2}'.format(
+            self.json_dir, self.name,
+            datetime.datetime.now().strftime("%Y%m%d-%H%M"))
         pytest.main([
             pytest_file, '-n={0}'.format(self.jobs), '-k={0}'.format(
                 self.filter), '--configlist={0}'.format(gen_config), '-v',
             '--seed={0}'.format(self.seed), '--count={0}'.format(self.count),
-            '--html={0}/aapg_gen.html'.format(output_dir),
+            '--html={0}/reports/aapg.html'.format(output_dir),
+            '--report-log={0}.json'.format(report_file_name),
             '--self-contained-html', '--output_dir={0}'.format(output_dir),
             '--module_dir={0}'.format(module_dir)
         ])
         # Generate Test List
         # Get the aapg dir from output
-        asm_dir = output_dir + 'aapg/asm/'
+        asm_dir = output_dir + '/aapg/asm/'
         test_list = {}
         asm_test_list = glob.glob(asm_dir + '**/*[!_template].S')
         # asm_templates = glob.glob(asm_dir+'/**/*.S')
@@ -117,6 +126,7 @@ class AapgPlugin(object):
                         isa.add('c')
             canonical_order = {'i': 0, 'm': 1, 'a': 2, 'f': 3, 'd': 4, 'c': 5}
             canonical_isa = sorted(list(isa), key=lambda d: canonical_order[d])
+
             march_str = 'rv' + str(xlen) + "".join(canonical_order)
             if xlen == 64:
                 mabi_str = 'lp64'
@@ -128,19 +138,25 @@ class AapgPlugin(object):
             base_key = os.path.basename(test)[:-2]
             test_list[base_key] = {}
             test_list[base_key][
-                'work_dir'] = output_dir + 'aapg/asm/' + base_key
+                'work_dir'] = output_dir + '/aapg/asm/' + base_key
             test_list[base_key]['isa'] = self.isa
             test_list[base_key]['march'] = march_str
             test_list[base_key]['mabi'] = mabi_str
             # test_list[base_key]['gcc_cmd'] = gcc_compile_bin + " " + "-march=" + arch + " " + "-mabi=" + abi + " " + gcc_compile_args + " -I " + asm_dir + include_dir + " -o $@ $< $(CRT_FILE) " + linker_args + " $(<D)/$*.ld"
             test_list[base_key]['cc'] = 'riscv64-unknown-elf-gcc'
             test_list[base_key][
-                'cc_args'] = '-march=' + march_str + ' -mabi=' + mabi_str + ' -mcmodel=medany -static -std=gnu99 -O2 -fno-common -fno-builtin-printf -I common'
+                'cc_args'] = ' -mcmodel=medany -static -std=gnu99 -O2 -fno-common -fno-builtin-printf -fvisibility=hidden '
             test_list[base_key][
                 'linker_args'] = '-static -nostdlib -nostartfiles -lm -lgcc -T'
-            test_list[base_key]['linker_file'] = base_key + '.ld'
-            test_list[base_key]['asm_file'] = base_key + '.S'
-            test_list[base_key]['crt_file'] = 'common/crt.S'
+            test_list[base_key][
+                'linker_file'] = output_dir + '/aapg/asm/' + base_key + '/' + base_key + '.ld'
+            test_list[base_key][
+                'asm_file'] = output_dir + '/aapg/asm/' + base_key + '/' + base_key + '.S'
+            test_list[base_key]['extra_compile'] = [
+                output_dir + '/aapg/common/crt.S'
+            ]
+            # TODO:DOC Add info possible results for the below variable
+            test_list[base_key]['result'] = 'Unavailable'
 
         return test_list
 
@@ -155,6 +171,8 @@ class AapgPlugin(object):
         """
         Overwrites the aapg entries in the regressfile with the latest present in the gendir
         """
+
+        output_dir = os.path.abspath(output_dir)
         remove_list = dict()
         test_dict['aapg']['aapg_global_testpath'] = output_dir
         if os.path.isdir(output_dir):
@@ -173,10 +191,11 @@ class AapgPlugin(object):
 
         if os.path.isfile(regressfile):
             with open(regressfile, 'r') as rgfile:
-                testlist = yaml.safe_load(rgfile)
+                testlist = utils.load_yaml(rgfile)
                 testlist['aapg'].update(test_dict)
             rgfile.close()
 
         rgfile = open(regressfile, 'w')
 
-        yaml.safe_dump(test_dict, rgfile, default_flow_style=False)
+        utils.yaml.dump(test_dict, rgfile)
+
